August 23th commit.
Made an activity so I can list all the reports from a line. Already did the adapter stuff and everything seems to be working. Next step is to work on the profile screen and on the notifications.

Made some adjustments on the database JSON file so I could retrieve data more easily.