August 26th commit.
Made the adjustments so the Facebook Auth works just fine again.

Created an splashscreen activity and a launcher to start the application. Rewrited the XML from the Login and Register activity cleaning some buttons and texts and putting then in the Launcher.

I did mess up with the branches and the August 23th commit isn't on this branch. On the next day, I'm probably going to merge them manually so I don't have any issues.