September 5th commit.
Remade all of the SettingsActivity. As I'm not using any preferences that I need to share, I changed everything to buttons and set listeners for then in the Java file, ez pz.
Already wrote the code for deleting the account and signing out. Need to add an Alert Dialog to the first one and figure out some kind of hyperlink to the About and Terms.

farewell